requirements.txt
prompt_toolkit
argparse
opencv-python
Pillow
numpy
pydicom